backend/requirements.txt
fastapi
feedparser
langchain
langchain-openai
langchain_community
langchain-ollama
langgraph
langsmith
pdfplumber 
openai
fastapi
uvicorn
pydantic
tiktoken
python-dotenv
aiohttp
requests
uvicorn